Update Tours carousels
